chore: No need to cast it to uint8

+++ test/foundry/TransferSelectorNFT.t.sol
@@ -10,21 +10,21 @@ contract TransferSelectorNFTTest is ProtocolBase, ITransferSelectorNFT {
     bytes4 newSelector = 0x69696969;
 
     function testInitialStates() public {
-        (address transferManager0, bytes4 selector0) = looksRareProtocol.managerSelectorOfAssetType(uint8(0));
+        (address transferManager0, bytes4 selector0) = looksRareProtocol.managerSelectorOfAssetType(0);
         assertEq(transferManager0, address(transferManager));
         assertEq(uint32(selector0), uint32(0xa7bc96d3));
 
-        (address transferManager1, bytes4 selector1) = looksRareProtocol.managerSelectorOfAssetType(uint8(1));
+        (address transferManager1, bytes4 selector1) = looksRareProtocol.managerSelectorOfAssetType(1);
         assertEq(transferManager1, address(transferManager));
         assertEq(uint32(selector1), uint32(0xa0a406c6));
     }
 
     function testAddTransferManagerForAssetType() public asPrankedUser(_owner) {
         vm.expectEmit(true, false, false, true);
-        emit NewAssetType(uint8(2), newTransferManager, newSelector);
-        looksRareProtocol.addTransferManagerForAssetType(uint8(2), newTransferManager, newSelector);
+        emit NewAssetType(2, newTransferManager, newSelector);
+        looksRareProtocol.addTransferManagerForAssetType(2, newTransferManager, newSelector);
 
-        (address transferManager, bytes4 selector) = looksRareProtocol.managerSelectorOfAssetType(uint8(2));
+        (address transferManager, bytes4 selector) = looksRareProtocol.managerSelectorOfAssetType(2);
         assertEq(transferManager, newTransferManager);
         assertEq(uint32(selector), uint32(newSelector));
     }
